+++ integration_test/multiport_test.go
@@ -144,6 +144,7 @@ func (s *connectedSuite) Test_UnnamedUdpAndTcpPort() {
 		}
 
 		itest.TelepresenceOk(ctx, "loglevel", "trace")
+		defer itest.TelepresenceOk(ctx, "loglevel", "debug")
 		itest.TelepresenceOk(ctx, "intercept", "-n", s.AppNamespace(), "--mount", "false", "--service", "echo-udp", "-p", fmt.Sprintf("%d:%s", localPort, svcPort), dep)
 		s.CapturePodLogs(ctx, "app="+dep, "traffic-agent", s.AppNamespace())
 		defer itest.TelepresenceOk(ctx, "leave", dep+"-"+s.AppNamespace())